requirements.txt
snowflake-sqlalchemy
panda
matplotlib
numpy
docopt
pymongo